lolapi/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
riotwatcher==3.2.4
django-cors-headers==3.13.0
django-environ==0.9.0
python-dotenv==0.21.0
mysqlclient==2.1.1